+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Feature
 * add `vox list` sub-command
 * add system integration with `vox install/remove`
+* add `vox at` command
+* add `vox config` command
 
 ## 0.2.0
 ### Feature